Move functions to utils.py
