+++ ads/llm/autogen/v02/loggers/session_logger.py
@@ -1,8 +1,6 @@
 # Copyright (c) 2024, Oracle and/or its affiliates.  All rights reserved.
 # Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl/
 import importlib
-import inspect
-import json
 import logging
 import os
 import tempfile
@@ -11,8 +9,7 @@ import traceback
 import uuid
 from dataclasses import dataclass, field
 from datetime import datetime, timedelta, timezone
-from types import SimpleNamespace
-from typing import Any, Dict, List, Optional, Tuple, Union
+from typing import Any, Dict, List, Optional, Union
 from urllib.parse import urlparse
 
 import autogen
@@ -44,85 +41,15 @@ from ads.llm.autogen.reports.data import (
 from ads.llm.autogen.reports.session import SessionReport
 from ads.llm.autogen.v02 import runtime_logging
 from ads.llm.autogen.v02.log_handlers.oci_file_handler import OCIFileHandler
+from ads.llm.autogen.v02.loggers.utils import (
+    serialize,
+    serialize_response,
+)
 
 logger = logging.getLogger(__name__)
 
 
 CONST_REPLY_FUNC_NAME = "reply_func_name"
-
-
-def is_json_serializable(obj: Any) -> bool:
-    """Checks if an object is JSON serializable.
-
-    Parameters
-    ----------
-    obj : Any
-        Any object.
-
-    Returns
-    -------
-    bool
-        True if the object is JSON serializable, otherwise False.
-    """
-    try:
-        json.dumps(obj)
-    except Exception:
-        return False
-    return True
-
-
-def serialize_response(response) -> dict:
-    """Serializes the LLM response to dictionary."""
-    if isinstance(response, SimpleNamespace) or is_json_serializable(response):
-        # Convert simpleNamespace to dict
-        return json.loads(json.dumps(response, default=vars))
-    elif hasattr(response, "dict") and callable(response.dict):
-        return json.loads(json.dumps(response.dict(), default=str))
-    data = {
-        "model": response.model,
-        "choices": [
-            {"message": {"content": choice.message.content}}
-            for choice in response.choices
-        ],
-        "response": str(response),
-    }
-    return data
-
-
-def serialize(
-    obj: Union[int, float, str, bool, Dict[Any, Any], List[Any], Tuple[Any, ...], Any],
-    exclude: Tuple[str, ...] = ("api_key", "__class__"),
-    no_recursive: Tuple[Any, ...] = (),
-) -> Any:
-    """Serializes an object for logging purpose."""
-    try:
-        if isinstance(obj, (int, float, str, bool)):
-            return obj
-        elif callable(obj):
-            return inspect.getsource(obj).strip()
-        elif isinstance(obj, dict):
-            return {
-                str(k): (
-                    serialize(str(v))
-                    if isinstance(v, no_recursive)
-                    else serialize(v, exclude, no_recursive)
-                )
-                for k, v in obj.items()
-                if k not in exclude
-            }
-        elif isinstance(obj, (list, tuple)):
-            return [
-                (
-                    serialize(str(v))
-                    if isinstance(v, no_recursive)
-                    else serialize(v, exclude, no_recursive)
-                )
-                for v in obj
-            ]
-        else:
-            return str(obj)
-    except Exception:
-        return str(obj)
 
 
 @dataclass
@@ -310,6 +237,11 @@ class SessionLogger(FileLogger):
     def log_file(self) -> Optional[str]:
         """Log file path for the current session."""
         return self.session.log_file
+
+    @property
+    def report(self) -> Optional[str]:
+        """Report path/link for the  session."""
+        return self.session.report
 
     @property
     def name(self) -> str:
